Scripts to install and test ArmoniK on windows using WSL2
These scripts allowed a fast installation of a full developper environment
on Windows, the manual step are limited to the wsl creation, the
username and password used and the name of the branch to deploy.

To do a deployment, in the working directory, start the powershell
script: armonik_dev_environment.ps1 and follow the instruction.